Stage 1 of 4 for JetBrains Academy - Kotlin -  Zookeeper project. Description
There are many animals in the zoo, and all of them need care. The animals must be fed, cleaned, surrounded by their kin, and kept happy. That is a difficult task for our large zoo, so one of your employers has suggested a more convenient way to keep track of everything. She wants to be able to pull up a video feed of any animal in the zoo with the help of a program. Being able to check on each habitat would help the zookeepers take care of our furry friends more efficiently!

In this project, you will create a program that helps the zookeepers check on the animals and make sure that they're doing well. Your product will be able to process commands from the zookeepers and display the animals on a monitor.
Objectives

To begin with, you will develop a simple printer. Your program should display the text from the output example.
Example

The output:

I love animals!
Let's check on the animals...
The deer looks fine.
The bat looks happy.
The lion looks healthy.

MainActivity.kt:
• added a println() to print the the required output.